院士列表返回数据实体类修改

+++ ruoyi-service/ruoyi-acad/src/main/java/com/ruoyi/acad/domain/Aos.java
@@ -32,7 +32,7 @@ public class Aos implements Serializable {
 
     @ApiModelProperty("科学院")
     @JsonSerialize(using = ToStringSerializer.class)
-    private Integer aosId;//科学院
+    private Long aosId;//科学院
 
     @ApiModelProperty("科学院名称")
     @NotNull
